\name{onlineFDR-deprecated}
\alias{onlineFDR-deprecated}
\title{Deprecated functions in package \sQuote{onlineFDR}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{onlineFDR} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{LORDdep: \code{\link{LORD}} with version='dep'}
    \item{bonfInfinite: \code{\link{AlphaSpending}}}

  }
}
