% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADDIS.R
\name{ADDIS}
\alias{ADDIS}
\title{ADDIS: Adaptive discarding algorithm for online FDR control}
\usage{
ADDIS(pval, alpha = 0.05, version, gammai, w0, lambda = 0.5,
  decision.times, lags, batch.sizes)
}
\arguments{
\item{pval}{A vector of p-values}

\item{alpha}{Overall significance level of the procedure, the default
is 0.05.}

\item{version}{Takes values 'async', 'dep' or 'batch'. This 
specifies the version of LORDstar to use.}

\item{gammai}{Optional vector of \eqn{\gamma_i}. A default is provided with
\eqn{\gamma_j} proportional to \eqn{1/j^2}.}

\item{w0}{Initial `wealth' of the procedure, defaults to \eqn{\alpha/10}.}

\item{lambda}{Optional threshold for a `candidate' hypothesis, must be 
between 0 and 1. Defaults to 0.5.}

\item{decision.times}{A vector of decision times for the hypothesis tests,
this is required for \code{version='async'}.}

\item{lags}{A vector of lags or the hypothesis tests, this is required for
\code{version='dep'}.}

\item{batch.sizes}{A vector of batch sizes, this is required for
\code{version='batch'}.}
}
\value{
\item{d.out}{A dataframe with the original pvalues \code{pval}, the
adjusted testing levels \eqn{\alpha_i} and the indicator 
function of discoveries \code{R}. Hypothesis \eqn{i} is rejected if the
\eqn{i}-th p-value is less than or equal to \eqn{\alpha_i}, in which case
\code{R[i] = 1}  (otherwise \code{R[i] = 0}).}
}
\description{
Implements the ADDIS algorithm for online FDR control, where ADDIS stands for
an ADaptive algorithm that DIScards conservative nulls, as presented by Tian
and Ramdas (2019).
}
\details{
The function takes as its input a vector of p-values, as well as a vector
describing the conflict sets for the hypotheses. This takes the form of a
vector of decision times, lags or batch sizes (see below).

Tian and Ramdas (2019). present explicit three versions of SAFFRONstar:

1) \code{version='async'} is for an asynchoronous testing process, consisting
of tests that start and finish at (potentially) random times. The discretised
finish times of the test correspond to the decision times. These decision
times are given as the input \code{decision.times} for this version of the
SAFFRONstar algorithm.

2) \code{version='dep'} is for online testing under local dependence of the
p-values. More precisely, for any \eqn{t>0} we allow the p-value \eqn{p_t}
have arbitrary dependence on the previous \eqn{L_t} p-values. The fixed
sequence of \eqn{L_t} is referred to as `lags' and is given as the input
\code{lags} for this version of the SAFFRONstar algorithm.

3) \code{version='batch'} is for controlling the mFDR in mini-batch testing,
where a mini-batch represents a grouping of tests run asynchronously which
result in dependent p-values. Once a mini-batch of tests is fully completed,
a new one can start, testing hypotheses independent of the previous batch.
The batch sizes are given as the input \code{batch.sizes} for this version of
the SAFFRONstar algorithm.

Given an overall significance level \eqn{\alpha}, SAFFRONstar depends on
constants \eqn{w_0} and \eqn{\lambda}, where \eqn{w_0} satisfies \eqn{0 \le
w_0 \le (1 - \lambda)\alpha} and represents the intial `wealth' of the
procedure, and \eqn{0 < \lambda < 1} represents the threshold for a
`candidate' hypothesis. A `candidate' refers to p-values smaller than
\eqn{\lambda}, since SAFFRON will never reject a p-value larger than
\eqn{\lambda}.

Note that these SAFFRONstar algorithms control the \emph{modified} FDR
(mFDR). The `async' version also controls the usual FDR if the p-values are
assumed to be independent.

Further details of the ADDIS algorithms can be found in Tian and Ramdas
(2019).
}
\examples{
pval = c(2.90e-08, 0.06743, 3.51e-04, 0.0174, 0.04723,
        3.60e-05, 0.79149, 0.27201, 0.28295, 7.59e-06,
        0.69274, 0.30443, 0.00136, 0.82342, 0.54757)


}
\references{
Tian, J. and Ramdas, A. (2019). ADDIS: an adaptive discarding algorithm for 
online FDR control with conservative nulls. \emph{arXiv preprint}, 
\url{https://arxiv.org/abs/1905.11465}.
}
\seealso{
\code{\link{SAFFRON}}
}
