% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LORDstar.R
\name{LORDstar}
\alias{LORDstar}
\title{LORDstar: Asychronous online mFDR control based on recent discovery}
\usage{
LORDstar(pval, alpha = 0.05, version, gammai, w0, decision.times, lags,
  batch.sizes)
}
\arguments{
\item{pval}{A vector of p-values}

\item{alpha}{Overall significance level of the procedure, the default
is 0.05.}

\item{version}{Takes values 'async', 'dep' or 'batch'. This 
specifies the version of LORDstar to use.}

\item{gammai}{Optional vector of \eqn{\gamma_i}. A default is provided as
proposed by Javanmard and Montanari (2018), equation 31.}

\item{w0}{Initial `wealth' of the procedure, defaults to \eqn{\alpha/10}.}

\item{decision.times}{A vector of decision times for the hypothesis tests,
this is required for \code{version='async'}.}

\item{lags}{A vector of lags or the hypothesis tests, this is required for
\code{version='dep'}.}

\item{batch.sizes}{A vector of batch sizes, this is required for
\code{version='batch'}.}
}
\value{
\item{d.out}{A dataframe with the original pvalues \code{pval}, the
adjusted testing levels \eqn{\alpha_i} and the indicator 
function of discoveries \code{R}. Hypothesis \eqn{i} is rejected if the
\eqn{i}-th p-value is less than or equal to \eqn{\alpha_i}, in which case
\code{R[i] = 1}  (otherwise \code{R[i] = 0}).}
}
\description{
Implements LORD algorithms for asynchronous online testing, as presented 
by Zrnic et al. (2018).
}
\details{
The function takes as its input a vector of p-values, as well as a vector
describing the conflict sets for the hypotheses. This takes the form of a
vector of decision times, lags or batch sizes (see below).

Zrnic et al. (2018) present explicit three versions of LORDstar:

1) \code{version='async'} is for an asynchoronous testing process,
consisting of tests that start and finish at (potentially) random times. The
discretised finish times of the test correspond to the decision times. These
decision times are given as the input \code{decision.times} for this version
of the LORDstar algorithm.

2) \code{version='dep'} is for online testing under local dependence
of the p-values. More precisely, for any \eqn{t>0} we allow the p-value
\eqn{p_t} have arbitrary dependence on the previous \eqn{L_t} p-values. The
fixed sequence of \eqn{L_t} is referred to as `lags' and is given as the
input \code{lags} for this version of the LORDstar algorithm.

3) \code{version='batch'} is for controlling the mFDR in mini-batch
testing, where a mini-batch represents a grouping of tests run 
asynchronously which result in dependent p-values. Once a mini-batch of tests
is fully completed, a new one can start, testing hypotheses independent of
the previous batch. The batch sizes are given as the input \code{batch.sizes}
for this version of the LORDstar algorithm.

Given an overall significance level \eqn{\alpha}, LORDstar depends on
\eqn{w_0} (where \eqn{0 \le w_0 \le \alpha}), which represents the intial
`wealth' of the procedure. The algorithms also require a sequence of
non-negative non-increasing numbers \eqn{\gamma_i} that sum to 1.

Note that these LORDstar algorithms control the \emph{modified} FDR (mFDR).
The `async' version also controls the usual FDR if the p-values are assumed
to be independent.

Further details of the LORDstar algorithms can be found in
Zrnic et al. (2018).
}
\examples{
pval = c(2.90e-08, 0.06743, 3.51e-04, 0.00174, 0.04723,
        3.60e-05, 0.79149, 0.27201, 0.28295, 7.59e-06,
        0.69274, 0.30443, 0.00136, 0.82342, 0.54757)

LORDstar(pval, version='async', decision.times=seq_len(15)) # Synchronous
LORDstar(pval, version='async', decision.times=seq_len(15)+1) # Asynchronous

LORDstar(pval, version='dep', lags=rep(0,15)) # Synchronous
LORDstar(pval, version='dep', lags=rep(1,15)) # Locally dependent

LORDstar(pval, version='batch', batch.sizes=rep(1,15)) # Synchronous
LORDstar(pval, version='batch', batch.sizes=c(4,6,5)) # Batched

}
\references{
Javanmard, A. and Montanari, A. (2018) Online Rules for Control of False
Discovery Rate and False Discovery Exceedance. \emph{Annals of Statistics},
46(2):526-554.

Zrnic et al. (2018). Asynchronous Online Testing of Multiple Hypotheses.
\emph{arXiv preprint}, \url{https://arxiv.org/abs/1812.05068}
}
\seealso{
\code{\link{LORD}} presents versions of LORD for \emph{synchronous} p-values,
i.e. where each test can only start when the previous test has finished.
}
