% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBound.R
\name{setBound}
\alias{setBound}
\title{setBound}
\usage{
setBound(alg, alpha = 0.05, N)
}
\arguments{
\item{alg}{A string that takes the value of one of the following: LOND, LORD, LORDdep, SAFFRON, ADDIS, LONDSTAR, LORDSTAR, SAFFRONSTAR, or alphainvesting}

\item{alpha}{Overall significance level of the FDR procedure, the default is 0.05.}

\item{N}{The number of hypotheses to be tested}
}
\value{
\item{bound}{A vector of nonnegative numbers that define the update rule to each subsequent significance threshold}
}
\description{
Sets an upper bound to the number of hypotheses to be tested
}
\details{
Note that when specifying the number of hypotheses to be tested, \code{N} should not be less than the number of hypotheses within your target dataset.
}
